pyproject.toml
[tool.poetry]
name = "in-delivery-text-generator"
version = "0.1.0"
description = ""
authors = ["Deverty <bychenkov.a.k@edu.mirea.ru>"]
readme = "README.md"
packages = [{ include = "in_delivery_text_generator" }]

[tool.poe.tasks]
run = "faststream run main:app"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.5"
ipykernel = "^6.27.0"
fastapi = "^0.104.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"
faststream = { extras = ["rabbit"], version = "^0.3.1" }
llama-cpp-python = "^0.2.20"
types-tqdm = "^4.66.0.5"
tqdm = "^4.66.1"
requests = "^2.31.0"
types-requests = "^2.31.0.10"
poethepoet = "^0.24.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
